server_name_generator/backend-app/app/requirements.txt
pip==18.1
bumpversion==0.5.3
wheel==0.32.1
watchdog==0.9.0
flake8==3.5.0
tox==3.5.2
coverage==4.5.1
Sphinx==1.8.1
twine==1.12.1
pytest==3.8.2
pytest-runner==4.2
Flask-Login==0.4.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.3.2
flask-ldap3-login==0.9.16
Flask-Cors==3.0.9
PyJWT==1.7.1
Flask-JWT-Extended==3.25.0
Werkzeug==2.0.1